Do not use plain object types for plain props.
It is easiest to explain this fix by example. Consider the following
schema:

```json
{
  "version": "0.0.1",
  "name": "example",
  "types": {
    "example::Foo": {
      "properties": {
        "a": {
          "type": "boolean"
        },
      },
      "type": "object"
    }
  },
  "resources": {
    "example::Component": {
      "isComponent": true,
      "properties": {
        "foo": {
          "$ref": "#/types/example::Foo"
        },
        "bar": {
          "$ref": "#/types/example::Foo"
        },
      },
      "inputProperties": {
        "foo": {
          "$ref": "#/types/example::Foo"
        },
        "bar": {
          "$ref": "#/types/example::Foo"
        },
      },
      "plainInputs": ["bar"],
      "type": "object"
    }
  },
}
```

This schema defines a single resource with two input properties, `foo`
and `bar`. The former is a normal input property; the latter is a plain
input property. When the code for this resource is generated, `foo`
should accept an inputty `example::Foo` (e.g. `Input<FooArgs>` in
NodeJS) and `bar` should accept a shallowly-unwrapped inputty
`example::Foo` (e.g. `FooArgs` in NodeJS).

Fixes #6957.